test: String split 수정, CarService 테스트 추가
